+++ src/pages/[artist].astro
@@ -1,17 +1,15 @@
 ---
-import type { Release } from "@types";
+import store from "@store";
 import Layout from "@layouts/Layout.astro";
-import collection from "@data/collection.json";
 import Releases from "@components/Releases.astro";
-import { getArtists, getReleasesByArtist } from "@util/data";
 
 export async function getStaticPaths() {
-  const allReleases = getReleasesByArtist(collection);
+  const allReleases = store.releasesByArtist;
 
   // TODO: There's likely a better way to do this than using getArtists since
   // we're already getting all the artists from getReleasesByArtist
   // We just don't have the artist slug there yet.
-  return getArtists(collection).map((artist) => {
+  return store.artists.map((artist) => {
     const artistReleases = allReleases[artist.name];
 
     return {
