+++ src/pages/artists.astro
@@ -1,9 +1,9 @@
 ---
+import store from "@store";
 import Layout from "@layouts/Layout.astro";
-import collection from "@data/collection.json";
-import { buildUrl, getArtists } from "@util/data";
+import { buildUrl } from "@util/data";
 
-const artists = getArtists(collection);
+const artists = store.artists;
 ---
 
 <Layout title="Artists">
