+++ src/pages/index.astro
@@ -1,11 +1,10 @@
 ---
+import type { Release } from "@types";
+import store from "@store";
 import Layout from "@layouts/Layout.astro";
 import Releases from "@components/Releases.astro";
-import collection from "@data/collection.json";
-import { getReleasesByArtist } from "@util/data";
-import type { Release } from "@types";
 
-const releases = getReleasesByArtist(collection);
+const releases = store.releasesByArtist;
 let releasesDisplay: Release[] = [];
 for (const key in releases) {
   releasesDisplay.push(...releases[key]);
