+++ src/types.ts
@@ -1,3 +1,10 @@
+interface SourceObject {
+  id: string | number;
+  [key: string]: any;
+}
+
+export type DataSource = SourceObject[];
+
 export interface Release {
   basic_information: any;
 }
